Added search for column to extract by checking for '3 months eneded' in th above it

+++ cpp/src/parser/Parser.h
@@ -146,7 +146,7 @@ private:
     bool extractBookValue(XmlElement* tree, DMMM::O_BalanceSheet& balance_data, string& units);
 // members - use to save relavent data for parsing
     DMMM::O_Stock _stock;
-    size_t _col_num;
+    int _col_num;
 };
 
 #endif //PARSER
